+++ lib/exceptions.dart
@@ -2,4 +2,6 @@
 library exceptions;
 
 export 'src/exceptions/byte_reader_exception.dart';
-export 'src/exceptions/socks_client_exception.dart';
+export 'src/exceptions/client/connection_closed_exception.dart';
+export 'src/exceptions/client/connection_command_failed_exception.dart';
+export 'src/exceptions/client/socks_client_exception.dart';