+++ lib/src/client/socks_client.dart
@@ -5,10 +5,12 @@ import 'dart:typed_data';
 import 'package:async/async.dart';
 import 'package:meta/meta.dart';
 
+import '../../exceptions.dart';
 import '../address_type.dart';
 import '../enums/authentication_method.dart';
 import '../enums/command_reply_code.dart';
 import '../enums/socks_connection_type.dart';
+import '../exceptions/client/connection_command_failed_exception.dart';
 import '../mixin/byte_reader.dart';
 import '../mixin/socket_mixin_.dart';
 import '../mixin/stream_mixin.dart';
@@ -64,22 +66,28 @@ class SocksSocket with StreamMixin<Uint8List>, SocketMixin, ByteReader {
     final socket = await Socket.connect(proxies.first.host, proxies.first.port);
   
     final client = SocksSocket.protected(socket, type);
-    await client._handshake(proxies.first);
 
-    for(var i = 1; i < proxies.length; i++) {
-      await client._handleCommand(proxies[i].host, proxies[i].port, SocksConnectionType.connect);
-      final response = await client._handleCommandResponse(SocksConnectionType.connect);
-      if(response.address != InternetAddress('0.0.0.0') || response.port != 0) {
-        throw UnimplementedError('Connect associated proxy not yet implemented.');
+    try {
+      await client._handshake(proxies.first);
+
+      for(var i = 1; i < proxies.length; i++) {
+        await client._handleCommand(proxies[i].host, proxies[i].port, SocksConnectionType.connect);
+        final response = await client._handleCommandResponse(SocksConnectionType.connect);
+        if(response.address != InternetAddress('0.0.0.0') || response.port != 0) {
+          throw UnimplementedError('Connect associated proxy not yet implemented.');
+        }
+        await client._handshake(proxies[i]);
       }
-      await client._handshake(proxies[i]);
+
+      await client._handleCommand(address, port, type);
+
+      final response = await client._handleCommandResponse(type);
+      
+      return SocksClientInitializeResult(client, response);
+    } on ByteReaderException catch (error, stackTrace) {
+      socket.close().ignore();
+      throw SocksClientConnectionClosedException((error: error, stackTrace: stackTrace));
     }
-
-    await client._handleCommand(address, port, type);
-
-    final response = await client._handleCommandResponse(type);
-
-    return SocksClientInitializeResult(client, response);
   }
 
   // Apply tls-over-http
@@ -199,9 +207,7 @@ class SocksSocket with StreamMixin<Uint8List>, SocketMixin, ByteReader {
  
     if (commandResponse != CommandReplyCode.succeed) {
       close().ignore();
-      throw Exception(
-        'Command handling failed. With error: ${commandResponse.name}',
-      );
+      throw SocksClientConnectionCommandFailedException(commandResponse);
     }
 
     // Read reserved byte.