+++ lib/src/exceptions/byte_reader_exception.dart
@@ -1,13 +1,8 @@
 import '../mixin/byte_reader.dart';
-import 'socks_client_exception.dart';
-
 
 /// Exception thrown by [ByteReader] mixin
-class ByteReaderException implements SocksClientException {
+class ByteReaderException implements Exception {
   const ByteReaderException(this.message);
 
   final String message;
-
-  @override
-  String toString() => message;
 }