+++ lib/src/exceptions/socks_client_exception.dart
@@ -1,7 +1,0 @@
-import '../client/socks_tcp_client.dart';
-import '../client/socks_udp_client.dart';
-
-/// Exception thrown by [SocksTCPClient] or [SocksUDPClient].
-class SocksClientException implements Exception {
-
-}